modules/controller/requirements.txt
azure-iot-device~=2.7.0
jsonschema
Adafruit_GPIO